fix(web): address code review issues #5–#10
5. IIFE in JSX (App.tsx): extract `const Tag: "div" | "main"` before the
   final return statement; removes the self-invoking function wrapper.

6. Unused _onOpenChat prop (HubDashboard): remove the prop from
   HubDashboardProps and from the HubMainContent call site — it was
   declared but never read inside the component.

7. pushAll missing offline queue on error (push.ts): move `modules`
   declaration before the try block so the catch handler can re-queue
   the payload, matching pushDirty's existing error path.

8. Keyboard DnD sensor (HubDashboard): add KeyboardSensor +
   sortableKeyboardCoordinates so module reordering is accessible via
   keyboard, not just pointer/touch.

9. Duplicate safeLS (insightsEngine.ts): replace the private JSON-parsing
   implementation with a one-liner that delegates to safeReadLS from
   @shared/lib/storage, eliminating the duplication.

10. Split useWeeklyDigest.ts (519 → 215 lines): extract all pure
    aggregation functions (aggregateFinyk/Fizruk/Nutrition/Routine) into
    core/lib/weeklyDigestAggregates.ts; re-export from useWeeklyDigest.ts
    so existing import paths in buildSlides.ts etc. are unaffected.

All 643 tests pass; TypeScript reports no errors.
